pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "cairn-tasks"
version = "0.1.0"
description = "Cairn - A minimal AI agent memory system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Bug Tracking",
]

[project.optional-dependencies]
mcp = ["fastmcp"]

[project.scripts]
cairn = "cairn:main"
